+++ spring-boot-2-demo-app/src/main/java/io/github/mfvanek/spring/boot2/test/controllers/HomeController.java
