+++ spring-boot-2-demo-app/src/test/java/io/github/mfvanek/spring/boot2/test/ApplicationTests.java
@@ -1,6 +1,6 @@
-package io.github.mfvanek.spring.test;
+package io.github.mfvanek.spring.boot2.test;
 
-import io.github.mfvanek.spring.test.support.TestBase;
+import io.github.mfvanek.spring.boot2.test.support.TestBase;
 import org.junit.jupiter.api.Test;
 
 class ApplicationTests extends TestBase {
