+++ spring-boot-2-demo-app/src/test/java/io/github/mfvanek/spring/boot2/test/support/KafkaConsumerUtils.java
@@ -1,4 +1,4 @@
-package io.github.mfvanek.spring.test.support;
+package io.github.mfvanek.spring.boot2.test.support;
 
 import lombok.experimental.UtilityClass;
 import org.apache.kafka.clients.CommonClientConfigs;
