+++ spring-boot-3-demo-app/src/main/java/io/github/mfvanek/spring/boot3/test/controllers/HomeController.java
@@ -1,4 +1,4 @@
-package io.github.mfvanek.spring.test.controllers;
+package io.github.mfvanek.spring.boot3.test.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
