+++ spring-boot-3-demo-app/src/test/java/io/github/mfvanek/spring/boot3/test/ApplicationTests.java
@@ -1,6 +1,6 @@
-package io.github.mfvanek.spring.test;
+package io.github.mfvanek.spring.boot3.test;
 
-import io.github.mfvanek.spring.test.support.TestBase;
+import io.github.mfvanek.spring.boot3.test.support.TestBase;
 import io.micrometer.observation.ObservationRegistry;
 import io.micrometer.tracing.Span;
 import io.micrometer.tracing.Tracer;
