+++ spring-boot-3-demo-app/src/test/java/io/github/mfvanek/spring/boot3/test/support/JaegerInitializer.java
@@ -1,4 +1,4 @@
-package io.github.mfvanek.spring.test.support;
+package io.github.mfvanek.spring.boot3.test.support;
 
 import org.springframework.boot.test.util.TestPropertyValues;
 import org.springframework.context.ApplicationContextInitializer;
